pyproject.toml
[tool.poetry]
name = "lesson20"
version = "0.1.0"
description = ""
authors = ["Artur Rubnikovich <arturredmu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
psycopg2 = "^2.9.9"
pillow = "^10.1.0"
django-ckeditor = "^6.7.0"
djangorestframework = "^3.14.0"
django-filter = "^23.5"


[tool.poetry.group.dev.dependencies]
faker = "^22.2.0"
django-debug-toolbar = "^4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
